+++ source/network.h
@@ -35,15 +35,15 @@ extern socklen_t g_sock_addr_len;
 
 class Network {
 public:
-	void set_inet_sock_addr(string, uint64_t, struct sockaddr_in&);
+	void set_inet_sock_addr(string, int, struct sockaddr_in&);
 	void bind_sock(int, struct sockaddr_in);
 	void get_sock_addr(int, struct sockaddr_in&);
 	void set_sock_reuse(int);
 	void set_rcv_timeout(int);
 	string get_src_ip_addr(Packet);
 	string get_dst_ip_addr(Packet);
-	void add_itf(uint64_t, string);
-	void rem_itf(uint64_t);
+	void add_itf(int, string);
+	void rem_itf(int);
 };
 
 extern Network g_nw;
