+++ source/packet.h
@@ -18,8 +18,8 @@ public:
 	S1ap s1ap_hdr;
 	Diameter diameter_hdr;
 	uint8_t *data;
-	uint64_t data_ptr;
-	uint64_t len;
+	int data_ptr;
+	int len;
 	
 	Packet();
 	Packet(const Packet&);
@@ -33,10 +33,10 @@ public:
 	void append_item(uint32_t);
 	void append_item(uint64_t);
 	void append_item(vector<uint64_t>);
-	void append_item(uint8_t*, uint64_t);	
+	void append_item(uint8_t*, int);	
 	void append_item(const char*);	
 	void append_item(string);
-	void prepend_item(uint8_t*, uint64_t);
+	void prepend_item(uint8_t*, int);
 	void prepend_gtp_hdr(uint8_t, uint8_t, uint16_t, uint32_t);
 	void prepend_s1ap_hdr(uint8_t, uint16_t, uint32_t, uint32_t);
 	void prepend_diameter_hdr(uint8_t, uint16_t);
@@ -46,16 +46,16 @@ public:
 	void extract_item(uint16_t&);
 	void extract_item(uint32_t&);
 	void extract_item(uint64_t&);
-	void extract_item(vector<uint64_t>&, uint64_t);
-	void extract_item(uint8_t*, uint64_t);	
-	void extract_item(char*, uint64_t);	
+	void extract_item(vector<uint64_t>&, int);
+	void extract_item(uint8_t*, int);	
+	void extract_item(char*, int);	
 	void extract_item(string&);
 	void extract_gtp_hdr();
 	void extract_s1ap_hdr();
 	void extract_diameter_hdr();
 	void truncate();
 	void clear_pkt();
-	struct ip* allocate_ip_hdr_mem(uint64_t);
+	struct ip* allocate_ip_hdr_mem(int);
 	~Packet();
 };
