+++ source/sctp_client.cpp
@@ -5,12 +5,12 @@ SctpClient::SctpClient() {
 	g_utils.handle_type1_error(conn_fd, "Socket error: sctpclient_sctpclient");
 }
 
-void SctpClient::conn(string arg_server_ip_addr, uint64_t arg_server_port) {
+void SctpClient::conn(string arg_server_ip_addr, int arg_server_port) {
 	init(arg_server_ip_addr, arg_server_port);
 	connect_with_server();
 }
 
-void SctpClient::init(string arg_server_ip_addr, uint64_t arg_server_port) {
+void SctpClient::init(string arg_server_ip_addr, int arg_server_port) {
 	int status;
 
 	server_port = arg_server_port;
